Move new files to correct packages
