+++ app/src/main/java/com/github/fribourgsdp/radio/database/FirestoreDatabase.kt
@@ -3,10 +3,7 @@ package com.github.fribourgsdp.radio.database
 import android.content.ContentValues
 import android.util.Log
 import com.github.fribourgsdp.radio.*
-import com.github.fribourgsdp.radio.data.Genre
-import com.github.fribourgsdp.radio.data.Playlist
-import com.github.fribourgsdp.radio.data.Song
-import com.github.fribourgsdp.radio.data.User
+import com.github.fribourgsdp.radio.data.*
 import com.github.fribourgsdp.radio.game.Game
 import com.google.android.gms.tasks.Task
 import com.google.android.gms.tasks.Tasks
