+++ app/src/main/java/com/github/fribourgsdp/radio/game/prep/JoinGameActivity.kt
@@ -13,8 +13,8 @@ import androidx.core.widget.addTextChangedListener
 import androidx.fragment.app.DialogFragment
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.github.fribourgsdp.radio.LobbyDataKeys
-import com.github.fribourgsdp.radio.PublicLobbiesAdapter
+import com.github.fribourgsdp.radio.data.LobbyDataKeys
+import com.github.fribourgsdp.radio.game.view.PublicLobbiesAdapter
 import com.github.fribourgsdp.radio.R
 import com.github.fribourgsdp.radio.config.MyAppCompatActivity
 import com.github.fribourgsdp.radio.data.User
